chore: refactor project-tests
