+++ gulpfile.ts
@@ -25,7 +25,6 @@ require('./scripts/gulp/tasks/docs');
 require('./scripts/gulp/tasks/screener');
 require('./scripts/gulp/tasks/stats');
 require('./scripts/gulp/tasks/test-unit');
-require('./scripts/gulp/tasks/test-projects');
 require('./scripts/gulp/tasks/perf');
 require('./scripts/gulp/tasks/test-e2e');
 require('./scripts/gulp/tasks/test-circulars');
