--- packages/fluentui/projects-test/.gulp.js
@@ -1,1 +1,0 @@
-module.exports = require('@uifabric/build/gulp/.gulp');
