+++ packages/fluentui/projects-test/assets/cra/App.tsx
