+++ packages/fluentui/projects-test/assets/nextjs/index.js
