+++ packages/fluentui/projects-test/assets/typings/index.tsx
