+++ packages/fluentui/projects-test/assets/typings/tsconfig.json
