+++ packages/fluentui/projects-test/babel.config.js
@@ -1,0 +1,4 @@
+module.exports = api => ({
+  ...require('@uifabric/build/babel')(api),
+  babelrcRoots: ['../../*'],
+});
