--- packages/fluentui/projects-test/gulpfile.ts
@@ -1,4 +1,0 @@
-import { series, task } from 'gulp';
-import '../../../gulpfile';
-
-task('test', series('test:projects'));
