+++ packages/fluentui/projects-test/package.json
@@ -7,14 +7,26 @@
     "@fluentui/react-northstar": "^0.51.0"
   },
   "devDependencies": {
+    "@lerna/project": "^3.15.0",
+    "@lerna/package-graph": "^3.14.0",
+    "@types/express": "^4.16.1",
+    "@types/fs-extra": "^8.0.1",
     "@uifabric/build": "^7.0.0",
+    "express": "^4.15.4",
     "gulp": "^4.0.2",
+    "fs-extra": "^8.1.0",
+    "portfinder": "^1.0.20",
+    "puppeteer": "^1.13.0",
+    "tmp": "^0.0.33",
     "typescript": "3.7.2"
   },
   "publishConfig": {
     "access": "public"
   },
   "scripts": {
-    "test": "gulp test"
+    "build": "tsc --noEmit",
+    "lint": "eslint --ext .js,.ts,.tsx .",
+    "lint:fix": "yarn lint --fix",
+    "test": "node -r @uifabric/build/babel/register src/index.ts"
   }
 }
